pathlib tests: significantly expand coverage of ABCs
1. Move all tests from `PurePosixPathTest` and `PureWindowsPathTest` into
   `PurePathTest`, and all tests from `PosixPathTest` and `WindowsPathTest`
   into `PathTest`. This ensures Posix- and Windows-specific tests run
   against `PurePath` and `Path`.
2. Move applicable tests from `PurePathTest` and `PathTest` to
   `PurePathBaseTest` and `PathBaseTest`. This increases coverage of Posix-
   and Windows-specific features in pathlib ABCs.
3. Add new `_raw.RawPurePath` and `RawPath` classes, which implement local
   paths without normalization. Move almost all remaining tests from
   `PurePathTest` and `PathTest` to `RawPurePathTest` and `RawPathTest`.
   This further increases coverage of Posix- and Windows-specific features
   in pathlib ABCs.
